pyproject.toml
[tool.pytest.ini_options]
pythonpath = ["."]

[tool.isort]
profile = "black"
line_length = 88
combine_as_imports = true

[tool.black]
line-length = 88
skip-magic-trailing-comma = true

[tool.bandit]
exclude_dirs = [".venv"]

[tool.poetry]
name = "discord_wireguard"
version = "0.0.0"
authors = ["Alex Massi"]
description = ""
readme = "README.md"
license = "AGPL-3"

[tool.poetry.dependencies]
python = "3.10.8"
hikari = "^2.0.0.dev115"
hikari-lightbulb = "^2.3.1"
wgconfig = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.12.0"
flake8 = "^6.0.0"
flake8-bugbear = "^22.12.6"
flake8-deprecated = "^2.0.1"
flake8-executable = "^2.1.2"
bandit = {extras = ["toml"], version = "^1.7.4"}
isort = "^5.11.4"
